+++ backEnd/src/main/java/com/example/Authentication/FormaPagamento/repository/FormaPagamentoRepository.java
@@ -8,5 +8,5 @@ import org.springframework.data.jpa.repository.Query;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface FormaPagamentoRepository extends JpaRepository<FormaPagamento, Short> {
+public interface FormaPagamentoRepository extends JpaRepository<FormaPagamento, Integer> {
 }
