+++ backEnd/src/main/java/com/example/Authentication/Fornecedores/repository/FornecedorRepository.java
@@ -9,7 +9,7 @@ import org.springframework.data.repository.query.Param;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface FornecedorRepository extends JpaRepository<Fornecedor, Short> {
+public interface FornecedorRepository extends JpaRepository<Fornecedor, Integer> {
     @Query(nativeQuery = true, value = "SELECT * FROM fornecedor p WHERE " +
             "(:search IS NULL OR :search = '' OR p.nome_razao_social LIKE %:search% OR p.cpf_cnpj " +
             "LIKE %:search% OR p.nome_fantasia LIKE %:search% OR p.telefone LIKE %:search%)")
