+++ backEnd/src/main/java/com/example/Authentication/Usuario/model/Usuario.java
@@ -21,7 +21,7 @@ import javax.validation.constraints.NotEmpty;
 public class Usuario {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Short id;
+    private Integer id;
     @Column
     @NotEmpty(message = "{campo.login.obrigatorio}")
     private String login;
