+++ backEnd/src/main/resources/db/migration/V16__CreateRelEspecialidadeAndMercadoria.sql
@@ -1,7 +1,7 @@
 CREATE TABLE rel_especialidades_mercadorias(
-    id SERIAL NOT NULL PRIMARY KEY,
-    id_especialidade SMALLINT NOT NULL,
-    id_mercadoria SMALLINT NOT NULL,
+    id INTEGER NOT NULL PRIMARY KEY,
+    id_especialidade INTEGER NOT NULL,
+    id_mercadoria INTEGER NOT NULL,
     FOREIGN KEY (id_especialidade) REFERENCES especialidade(id) ON UPDATE CASCADE ON DELETE CASCADE,
     FOREIGN KEY (id_mercadoria) REFERENCES mercadoria(id) ON UPDATE CASCADE ON DELETE CASCADE
 );
