+++ backEnd/src/main/resources/db/migration/V20__DropTable.sql
@@ -1,10 +1,10 @@
 DROP TABLE rel_especialidades_mercadorias;
 
 CREATE TABLE especialidade_mercadoria(
-    id SERIAL NOT NULL PRIMARY KEY,
-    id_especialidade SMALLINT NOT NULL,
-    id_mercadoria SMALLINT NOT NULL,
-    quantidade SMALLINT,
+    id INTEGER NOT NULL PRIMARY KEY,
+    id_especialidade INTEGER NOT NULL,
+    id_mercadoria INTEGER NOT NULL,
+    quantidade INTEGER,
     FOREIGN KEY (id_especialidade) REFERENCES especialidade(id) ON UPDATE CASCADE ON DELETE CASCADE,
     FOREIGN KEY (id_mercadoria) REFERENCES mercadoria(id) ON UPDATE CASCADE ON DELETE CASCADE
 );
