+++ backEnd/src/main/resources/db/migration/V21__CreateTableCaixa.sql
@@ -10,23 +10,23 @@ CREATE TABLE caixa (
 );
 
 CREATE TABLE pedido (
-  id SMALLINT NOT NULL AUTO_INCREMENT,
+  id INTEGER NOT NULL AUTO_INCREMENT,
   nome_cliente VARCHAR(100),
   observacao TEXT,
-  mesa SMALLINT,
+  mesa INTEGER,
   ativo INTEGER,
   valor_total DOUBLE,
   id_caixa INTEGER NOT NULL,
-  id_forma_pagamento SMALLINT,
+  id_forma_pagamento INTEGER,
   FOREIGN KEY (id_caixa) REFERENCES caixa(id) ON UPDATE CASCADE ON DELETE CASCADE,
   FOREIGN KEY (id_forma_pagamento) REFERENCES forma_pagamento(id) ON UPDATE CASCADE ON DELETE CASCADE,
   PRIMARY KEY (id)
 );
 
 CREATE TABLE pedidos_mercadoria (
-  id SMALLINT NOT NULL AUTO_INCREMENT,
-  id_mercadoria SMALLINT NOT NULL,
-  id_pedido SMALLINT NOT NULL,
+  id INTEGER NOT NULL AUTO_INCREMENT,
+  id_mercadoria INTEGER NOT NULL,
+  id_pedido INTEGER NOT NULL,
   quantidade INTEGER,
   FOREIGN KEY (id_pedido) REFERENCES pedido(id) ON UPDATE CASCADE ON DELETE CASCADE,
   FOREIGN KEY (id_mercadoria) REFERENCES mercadoria(id) ON UPDATE CASCADE ON DELETE CASCADE,
@@ -34,9 +34,9 @@ CREATE TABLE pedidos_mercadoria (
 );
 
 CREATE TABLE pedidos_especialidade (
-  id SMALLINT NOT NULL AUTO_INCREMENT,
-  id_especialidade SMALLINT NOT NULL,
-  id_pedido SMALLINT NOT NULL,
+  id INTEGER NOT NULL AUTO_INCREMENT,
+  id_especialidade INTEGER NOT NULL,
+  id_pedido INTEGER NOT NULL,
   quantidade INTEGER,
   FOREIGN KEY (id_pedido) REFERENCES pedido(id) ON UPDATE CASCADE ON DELETE CASCADE,
   FOREIGN KEY (id_especialidade) REFERENCES especialidade(id) ON UPDATE CASCADE ON DELETE CASCADE,
@@ -44,9 +44,9 @@ CREATE TABLE pedidos_especialidade (
 );
 
 CREATE TABLE adicional_pedido_especialidade (
-  id SMALLINT NOT NULL AUTO_INCREMENT,
-  id_mercadoria SMALLINT NOT NULL,
-  id_pedidos_especialidade SMALLINT NOT NULL,
+  id INTEGER NOT NULL AUTO_INCREMENT,
+  id_mercadoria INTEGER NOT NULL,
+  id_pedidos_especialidade INTEGER NOT NULL,
   quantidade INTEGER,
   FOREIGN KEY (id_pedidos_especialidade) REFERENCES pedidos_especialidade(id) ON UPDATE CASCADE ON DELETE CASCADE,
   FOREIGN KEY (id_mercadoria) REFERENCES mercadoria(id) ON UPDATE CASCADE ON DELETE CASCADE,
